separated classes and tpls
