+++ include/SuiteEditor/tpls/SuiteEditorConnector.tpl
@@ -1,0 +1,2 @@
+{literal}<script>if(!SuiteEditor) {var SuiteEditor = {};}</script>{/literal}
+{$editor}
